requirements.txt
# FastAPI and web framework
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
starlette==0.37.2
jinja2==3.1.3
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration and validation
pydantic==2.6.4
pydantic-settings==2.1.0
python-dotenv==1.0.1

# AWS services
boto3==1.34.80
botocore==1.34.80

# Authentication
cognitojwt==1.4.1

# CCAvenue Payment Gateway
pay-ccavenue>=1.0.0
# Note: pycryptodome may still be needed as a dependency for pay-ccavenue
pycryptodome==3.20.0

# HTTP client
requests==2.31.0
httpx==0.26.0

# Utilities
python-dateutil==2.9.0.post0
pyyaml==6.0.1

# Logging
structlog==24.1.0
